+++ src/main/scala/webmodelica/controllers/InfoController.scala
@@ -11,30 +11,8 @@ package webmodelica.controllers
 import com.twitter.finagle.http.Request
 import com.twitter.finatra.http.Controller
 import com.google.inject.Inject
+import webmodelica.models.Infos
 import webmodelica.models.config._
-import io.circe.generic.JsonCodec
-
-@JsonCodec
-case class Infos(
-  appName:String,
-  version:String,
-  copyright:String,
-  license:String,
-  licenseUri:String,
-  commitHash:String,
-)
-object Infos {
-  import buildinfo.BuildInfo
-
-  def apply():Infos = Infos(
-    BuildInfo.name,
-    BuildInfo.version,
-    BuildInfo.copyright,
-    BuildInfo.license,
-    BuildInfo.licenseUri,
-    BuildInfo.commit
-  )
-}
 
 class InfoController @Inject()(config: WMConfig, prefix:webmodelica.ApiPrefix)
     extends Controller {
